+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/JacocoDeltaReportGeneratorFacade.kt
@@ -1,7 +1,8 @@
 package io.github.surpsg.deltacoverage.report.jacoco
 
-import io.github.surpsg.deltacoverage.report.CoverageVerificationResult
+import io.github.surpsg.deltacoverage.report.CoverageSummary
 import io.github.surpsg.deltacoverage.report.DeltaReportGeneratorFacade
+import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportContext
 import io.github.surpsg.deltacoverage.report.jacoco.analyzable.AnalyzableReport
 import io.github.surpsg.deltacoverage.report.jacoco.analyzable.analyzableReportFactory
@@ -18,21 +19,26 @@ import org.jacoco.report.MultiSourceFileLocator
 
 internal class JacocoDeltaReportGeneratorFacade : DeltaReportGeneratorFacade() {
 
-    override fun generate(reportContext: ReportContext): List<CoverageVerificationResult> {
+    override fun generate(reportContext: ReportContext): CoverageSummary {
         val analyzableReports: Set<AnalyzableReport> = analyzableReportFactory(reportContext)
 
         val execFileLoader = CoverageLoader.loadExecFiles(reportContext.binaryCoverageFiles)
 
-        return analyzableReports.flatMap {
-            create(reportContext, execFileLoader, it)
-        }
+        val summaries: Map<ReportBound, CoverageSummary> = analyzableReports
+            .asSequence()
+            .map { it.reportBound to it }
+            .map { (reportBound, analyzableReport) ->
+                reportBound to create(reportContext, execFileLoader, analyzableReport)
+            }
+            .toMap()
+        return summaries.getValue(ReportBound.DELTA_REPORT)
     }
 
     private fun create(
         reportContext: ReportContext,
         execFileLoader: ExecFileLoader,
         analyzableReport: AnalyzableReport,
-    ): List<CoverageVerificationResult> {
+    ): CoverageSummary {
         val bundleCoverage: IBundleCoverage = analyzeStructure(reportContext) { coverageVisitor ->
             analyzableReport.buildAnalyzer(execFileLoader.executionDataStore, coverageVisitor)
         }
@@ -51,7 +57,13 @@ internal class JacocoDeltaReportGeneratorFacade : DeltaReportGeneratorFacade() {
 
             visitEnd()
         }
-        return verifiableVisitor.verificationResults
+        return CoverageSummary(
+            reportBound = verifiableVisitor.reportBound,
+            view = reportContext.deltaCoverageConfig.view,
+            coverageRulesConfig = reportContext.deltaCoverageConfig.coverageRulesConfig,
+            verifications = verifiableVisitor.verificationResults,
+            coverageInfo = verifiableVisitor.coverageInfo
+        )
     }
 
     private fun analyzeStructure(