+++ delta-coverage-core/src/test/kotlin/io/github/surpsg/deltacoverage/report/CoverageViolationsPropagatorTest.kt
@@ -1,7 +1,11 @@
 package io.github.surpsg.deltacoverage.report
 
+import io.github.surpsg.deltacoverage.config.CoverageEntity
 import io.github.surpsg.deltacoverage.config.CoverageRulesConfig
 import io.github.surpsg.deltacoverage.exception.CoverageViolatedException
+import io.github.surpsg.deltacoverage.report.CoverageSummary.Info
+import io.github.surpsg.deltacoverage.report.CoverageSummary.VerificationResult
+import io.github.surpsg.deltacoverage.report.CoverageViolationsPropagatorTest.CoverageSummaryBuilder.Companion.coverageSummary
 import io.kotest.assertions.throwables.shouldNotThrow
 import io.kotest.assertions.throwables.shouldThrow
 import org.junit.jupiter.api.Test
@@ -12,57 +16,99 @@ class CoverageViolationsPropagatorTest {
     @Test
     fun `propagate should throw CoverageViolatedException if failOnViolation is true and there are violations`() {
         // GIVEN
-        val verificationResults = listOf(
-            CoverageVerificationResult(
-                "any view",
-                CoverageRulesConfig {
-                    failOnViolation = true
-                },
-                listOf("violation1", "violation2")
+        val coverageSummary = coverageSummary {
+            coverageRulesConfig = CoverageRulesConfig {
+                failOnViolation = true
+            }
+            verifications += VerificationResult(
+                coverageEntity = CoverageEntity.INSTRUCTION,
+                violation = "any violation 3"
             )
-        )
+        }
 
         // WHEN // THEN
         shouldThrow<CoverageViolatedException> {
-            propagator.propagateAll(verificationResults)
+            propagator.propagateAll(coverageSummary)
         }
     }
 
     @Test
     fun `propagate should not throw any exception if failOnViolation is true and there are no violations`() {
         // GIVEN
-        val verificationResults = listOf(
-            CoverageVerificationResult(
-                "any view",
-                CoverageRulesConfig {
-                    failOnViolation = true
-                },
-                emptyList()
-            )
-        )
+        val coverageSummary = coverageSummary {
+            coverageRulesConfig = CoverageRulesConfig {
+                failOnViolation = true
+            }
+            verifications.clear()
+        }
 
         // WHEN // THEN
         shouldNotThrow<CoverageViolatedException> {
-            propagator.propagateAll(verificationResults)
+            propagator.propagateAll(coverageSummary)
         }
     }
 
     @Test
     fun `propagate should not throw any exception if failOnViolation is false`() {
         // GIVEN
-        val verificationResults = listOf(
-            CoverageVerificationResult(
-                "any view",
-                CoverageRulesConfig {
-                    failOnViolation = false
-                },
-                listOf("violation1", "violation2")
+        val coverageSummary = coverageSummary {
+            coverageRulesConfig = CoverageRulesConfig {
+                failOnViolation = false
+            }
+            verifications += VerificationResult(
+                coverageEntity = CoverageEntity.INSTRUCTION,
+                violation = "any violation 3"
             )
-        )
+        }
 
         // WHEN // THEN
         shouldNotThrow<CoverageViolatedException> {
-            propagator.propagateAll(verificationResults)
+            propagator.propagateAll(coverageSummary)
+        }
+    }
+
+    private class CoverageSummaryBuilder {
+        var view: String = "any view"
+        var coverageRulesConfig: CoverageRulesConfig = CoverageRulesConfig {
+            failOnViolation = false
+        }
+        var verifications: MutableList<VerificationResult> = mutableListOf(
+            VerificationResult(
+                coverageEntity = CoverageEntity.LINE,
+                violation = "any violation 1"
+            ),
+            VerificationResult(
+                coverageEntity = CoverageEntity.BRANCH,
+                violation = "any violation 2"
+            )
+        )
+        var coverageInfo: MutableList<Info> = mutableListOf(
+            Info(
+                coverageEntity = CoverageEntity.LINE,
+                covered = 1,
+                total = 2
+            ),
+            Info(
+                coverageEntity = CoverageEntity.BRANCH,
+                covered = 3,
+                total = 4
+            )
+        )
+
+        private fun build() = CoverageSummary(
+            reportBound = ReportBound.DELTA_REPORT,
+            view = view,
+            coverageRulesConfig = coverageRulesConfig,
+            verifications = verifications,
+            coverageInfo = coverageInfo,
+        )
+
+        companion object {
+            fun coverageSummary(
+                customize: CoverageSummaryBuilder.() -> Unit = {}
+            ): CoverageSummary {
+                return CoverageSummaryBuilder().apply(customize).build()
+            }
         }
     }
 }