+++ delta-coverage-gradle/src/functionalTest/kotlin/io/github/surpsg/deltacoverage/gradle/DeltaCoverageGradleReleasesTest.kt
@@ -50,6 +50,5 @@ class DeltaCoverageGradleReleasesTest {
         gradleRunner
             .withGradleVersion(gradleVersion)
             .runDeltaCoverageTask()
-            .assertDeltaCoverageStatusEqualsTo(SUCCESS)
     }
 }